+++ src/audio/midi_player_params.rs
@@ -1,12 +1,12 @@
 /// Hold values changed during playback of a MIDI events.
 pub struct MidiPlayerParams {
-    tempo: i32,
-    tempo_percentage: usize,
+    tempo: u32,
+    tempo_percentage: u32,
     solo_track_id: Option<usize>,
 }
 
 impl MidiPlayerParams {
-    pub const fn new(tempo: i32, tempo_percentage: usize, solo_track_id: Option<usize>) -> Self {
+    pub const fn new(tempo: u32, tempo_percentage: u32, solo_track_id: Option<usize>) -> Self {
         Self {
             tempo,
             tempo_percentage,
@@ -22,15 +22,15 @@ impl MidiPlayerParams {
         self.solo_track_id = solo_track_id;
     }
 
-    pub fn adjusted_tempo(&self) -> i32 {
-        (self.tempo as f32 * self.tempo_percentage as f32 / 100.0) as i32
+    pub fn adjusted_tempo(&self) -> u32 {
+        (self.tempo as f32 * self.tempo_percentage as f32 / 100.0) as u32
     }
 
-    pub fn set_tempo(&mut self, tempo: i32) {
+    pub fn set_tempo(&mut self, tempo: u32) {
         self.tempo = tempo;
     }
 
-    pub fn set_tempo_percentage(&mut self, tempo_percentage: usize) {
+    pub fn set_tempo_percentage(&mut self, tempo_percentage: u32) {
         self.tempo_percentage = tempo_percentage;
     }
 }
